# openapi.yaml
openapi: 3.0.3
info:
  title: SGS - Système de Gestion Scolaire API
  description: |
    API RESTful pour le Système de Gestion Scolaire (SGS).
    Cette spécification définit tous les endpoints nécessaires pour le backend.
    
    ## Authentification
    L'API utilise JWT Bearer tokens pour l'authentification.
    Tous les endpoints (sauf /auth/login) requièrent un header Authorization valide.
    
    ## Pagination
    Les endpoints de liste supportent la pagination avec les paramètres:
    - `page`: numéro de page (défaut: 1)
    - `per_page`: éléments par page (défaut: 20, max: 100)
    - `sort_by`: champ de tri
    - `sort_order`: asc ou desc
    - `search`: recherche textuelle
    
    ## Codes d'erreur
    - 400: Requête invalide
    - 401: Non authentifié
    - 403: Accès refusé
    - 404: Ressource non trouvée
    - 422: Erreur de validation
    - 500: Erreur serveur
  version: 1.0.0
  contact:
    name: Équipe SGS
    email: dev@sgs.edu
servers:
  - url: http://localhost:3001/api
    description: Serveur de développement
  - url: https://api.sgs.edu
    description: Serveur de production
...
paths:
...
  /rh/presences:
    get:
      tags: [RH]
      summary: Liste des présences
      security:
        - bearerAuth: []
      parameters:
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: employe_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Liste des présences
    post:
      tags: [RH]
      summary: Enregistrer une présence
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PresenceCreate'
      responses:
        '201':
          description: Présence enregistrée
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    PresenceCreate:
      type: object
      required: [employe_id, employe_type, date, statut]
      properties:
        employe_id:
          type: string
        employe_type:
          type: string
          enum: [enseignant, personnel]
        date:
          type: string
          format: date
        heure_arrivee:
          type: string
          format: time
        heure_depart:
          type: string
          format: time
        statut:
          type: string
          enum: [present, absent, retard, conge, mission]
        motif:
          type: string
...